Restore C demo to drive. Please add a weights bin
